Optimize library size by inlining graph logic
