+++ src/index.js
@@ -1,23 +1,51 @@
-import Graph from './graph';
-
 const Topologica = options => {
   const values = {};
   const functions = {};
-  const graph = Graph();
+  const edges = {};
+  const adjacent = node => edges[node] || [];
+
+  const addNode = node => {
+    edges[node] = adjacent(node);
+  };
+
+  const addEdge = (u, v) => {
+    addNode(u);
+    addNode(v);
+    adjacent(u).push(v);
+  };
+
+  const depthFirstSearch = sourceNodes => {
+    const visited = {};
+    const nodeList = [];
+
+    const DFSVisit = node => {
+      if (!visited[node]) {
+        visit(node);
+        nodeList.push(node);
+      }
+    };
+
+    const visit = node => {
+      visited[node] = true;
+      adjacent(node).forEach(DFSVisit);
+    }
+
+    sourceNodes.forEach(visit);
+
+    return nodeList;
+  }
 
   const invoke = property => {
     functions[property]();
   };
 
   const set = function(options) {
-    graph
-      .topologicalSort(Object.keys(options).map(property => {
-        if (values[property] !== options[property]) {
-          values[property] = options[property];
-          return property;
-        }
-      }))
-      .forEach(invoke);
+    depthFirstSearch(Object.keys(options).map(property => {
+      if (values[property] !== options[property]) {
+        values[property] = options[property];
+        return property;
+      }
+    })).reverse().forEach(invoke);
     return this;
   };
 
@@ -42,7 +70,7 @@ const Topologica = options => {
       : dependencies;
 
     dependencies.forEach(input => {
-      graph.addEdge(input, property);
+      addEdge(input, property);
     });
 
     functions[property] = () => {
